+++ app/Exceptions/Store/PaymentProcessorException.php
@@ -5,10 +5,8 @@
 
 declare(strict_types=1);
 
-namespace App\Libraries\Fulfillments;
+namespace App\Exceptions\Store;
 
-use App\Exceptions\Store\OrderException;
-
-class FulfillmentException extends OrderException
+class PaymentProcessorException extends OrderException
 {
 }
