+++ app/Libraries/Fulfillments/OrderFulfiller.php
@@ -6,6 +6,7 @@
 namespace App\Libraries\Fulfillments;
 
 use App\Events\Fulfillments\FulfillmentValidationFailed;
+use App\Exceptions\Store\FulfillmentException;
 use App\Models\Store\Order;
 use App\Traits\Validatable;
